data.py -- added dataset number to toHDUl, also added comments
Now use EXTNAME keyword to name the extensions which helps when using 'fv' to examine a FITS file. Also now check for a version number in the header and if it is not present comments explaining the data format are added. This should avoid adding the comments in multiple times.