doppler.cc, comdat.py -- bug fix, abs to fabs
Chris Watson & Ernst de Mooij were having a problem with NaNs. I traced it to the incorrect use of "abs" (returns an integer) as opposed to "fabs". This might have affected other computations in a more subtle way but its hard for me to tell. It is used to divide by the velocity range, in effect to convert to a "flux density" (flux per km/s). The main question is whether the velocity range varies. I think the best approach is to regard previous computations as dodgy at best, even if they did not turn up NaNs. The reason it crashed for the QUB guys was they had a velocity range < 1 km/s which was then set to zero by the use of abs.